refactor(router-store): use env initializer for connecting service
