+++ modules/router-store/src/index.ts
@@ -22,15 +22,15 @@ export {
   routerRequestAction,
 } from './actions';
 export { routerReducer, RouterReducerState } from './reducer';
+export { StoreRouterConnectingModule } from './router_store_module';
 export {
   StateKeyOrSelector,
-  StoreRouterConnectingModule,
   StoreRouterConfig,
   NavigationActionTiming,
   ROUTER_CONFIG,
   DEFAULT_ROUTER_FEATURENAME,
   RouterState,
-} from './router_store_module';
+} from './router_store_config';
 export {
   RouterStateSerializer,
   BaseRouterStoreState,
