+++ modules/router-store/src/router_selectors.ts
@@ -5,7 +5,7 @@ import {
 } from '@ngrx/store';
 import { RouterStateSelectors } from './models';
 import { RouterReducerState } from './reducer';
-import { DEFAULT_ROUTER_FEATURENAME } from './router_store_module';
+import { DEFAULT_ROUTER_FEATURENAME } from './router_store_config';
 
 export function createRouterSelector<
   State extends Record<string, any>
